refactor(sdk): keep AuthConsent as a component beside the sign-up view
Restores the consent checkbox as its own component, placed in the
sign-up folder and imported only from there, instead of inlined into
the view.
