+++ web/sdk/client/views/auth/sign-up/auth-consent.module.css
@@ -1,0 +1,3 @@
+.container {
+  width: 100%;
+}
