+++ web/sdk/client/views/auth/sign-up/sign-up-view.module.css
@@ -1,7 +1,3 @@
 .redirectLink {
   color: var(--rs-color-foreground-accent-primary);
 }
-
-.consent {
-  width: 100%;
-}
